app/requirements.txt
accelerate==0.29.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
psutil==5.9.8
PyPDF2==3.0.1
pytesseract==0.3.10
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
